FDMThetaMethod requires not more than FiniteDifference1DModel.
Also: getNumberOfStandardDeviations is a double (not an int).

+++ src/main/java/net/finmath/finitedifference/models/FDMBlackScholesModel.java
@@ -83,7 +83,7 @@ public class FDMBlackScholesModel implements FiniteDifference1DModel {
 		return numSpacesteps;
 	}
 
-	public int getNumStandardDeviations() {
+	public double getNumStandardDeviations() {
 		return numStandardDeviations;
 	}
